[DotNet] Add signed narrow scalars
Map Kotlin Byte and Short to exact CLR int8 and int16 carriers while preserving Common promotion, conversion, nullable, boxing, generic, and overload semantics.

Normalize boxed hash and invariant string behavior across Framework CLR and CoreCLR, and accept matching foreign SByte and Int16 signatures. Freeze the ABI with adversarial compiler, IL, cross-module, and C# product tests on both profiles.

+++ compiler/testData/codegen/dotnet/ilText/nestedClassesRejected.kt
@@ -1,5 +1,5 @@
 // Unsupported nested shapes remove only their own metadata subtree. BrokenFamily reaches the
-// member pre-pass before Bad's Byte signature fails; Good, the root, and UsesBroken survive.
+// member pre-pass before Bad's Float signature fails; Good, the root, and UsesBroken survive.
 // DeepBrokenFamily instead fails while recursively rendering a three-level nested member body,
 // exercising deepest-failure attribution while its valid metadata ancestors survive.
 // BrokenNestedBaseFamily pins the inheritance cascade: its nested base fails the member pre-pass,
@@ -19,7 +19,7 @@ class GenericInnerHost<T> {
 
 class BrokenInnerHost {
     inner class Bad {
-        fun unsupported(value: Byte): Byte = value
+        fun unsupported(value: Float): Float = value
     }
 
     class Good {
@@ -47,7 +47,7 @@ class BrokenFamily {
     }
 
     class Bad {
-        fun unsupported(value: Byte): Byte = value
+        fun unsupported(value: Float): Float = value
     }
 }
 
@@ -57,7 +57,7 @@ class UsesBroken {
 
 class BrokenNestedBaseFamily {
     open class Base {
-        fun unsupported(value: Byte): Byte = value
+        fun unsupported(value: Float): Float = value
     }
 }
